Add array version of cut!
This version allows cutting on arrays of traces either with
constants, or with arrays specifying where to cut for each
SACtr element.  This allows one to do things like:

````julia
cut!(S, S[:t0] - 100, S[:t1] + 150)
```

That is, one can cut traces to different times and lengths,
for example by the time picks in the different headers.
Obviously, arbitrary arrays of pick times can be used.