Add routines which return copies of traces
Routines `bandpass!()` and so on have equivalents like `bandpass()`
which return a copy of the trace(s) passed in.  Mostly they are
generated by metaprogramming, so new methods should be written
for in-place operation.  An exception is where pairs of traces
are expected (e.g., `rotate_through!()` which aren't dealt with
by the metaprogramming yet.