WIP: Allow function chaining; always import Dierckx
Experimental support for function chaining, where one does

    s::SACtr |> rtrend |> taper |> bp(0.1, 2) |> cut(:t0, -10, :t1, 30)

This is achieved by adding methods for most functions which accept a
SACtr of Array{SACtr} as the first argument and returning anonymous
functions which apply the methods to the output of the previous
command in the chain.  This can make some processing look nicer.

Note that this is experimental and may be removed.